Repeat of tenstorrent/tt-smi#191

> ASIC temperature in BH telemetry is a 16.16 fixed point integer signed
using two's complement. Negative values weren't being computed and so
were being displayed as 65536 - temp.

+++ crates/luwen-api/src/chip/mod.rs
@@ -257,7 +257,7 @@ impl Telemetry {
             let frac: f64 = (self.asic_temperature & 0xFFFF).into();
             let frac = frac / 65536.0;
 
-            let int: f64 = (self.asic_temperature >> 16).into();
+            let int: f64 = ((self.asic_temperature >> 16) as i16).into();
 
             int + frac
         } else {
